Rework --pg-upgrade-jobs and generalize to --jobs
We are generalizing the --pg-upgrade-jobs flag
in preparation for utilizing it to specify the
number of connections to use when executing
data migration script SQL statements in parallel.

As a result, --jobs can now be set for data migration
subcommands, initialize, finalize, and revert.

This commit also regenerates the protocol buffers.