Add NewPool function
This commit introduces the concept of a connection Pool to
execute SQL statements in parallel. We use the pgxpool package
to do so, and update the pgx package to v5 as well.

NewPool is similar to the Cluster.Connection method in
that it accepts a set of Options, but instead of returning a
connection string, it returns a pgxpool.Pool object that can be
used to execute queries against the specified database.

The maximum active connection limit is set when establishing the pool
by passing a value to NumConns():

	pool, err := greenplum.NewConnectionPool(greenplum.Port(port),
             greenplum.Database(dbName), greenplum.NumConns(numConns))
	if err != nil {
		return err
	}
	defer pool.Close()